refactor(frontend/user): add user module
