+++ packages/frontend/src/App.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import './styles/app.css';
-import { Main, User, UserProducts, UserCart } from './components/pages/';
+import { Main } from './components/pages/';
 import Login from './components/modules/Login/Login';
 import Register from './components/modules/Register/Register';
 import Home from './components/modules/Home/Home';
 import Payment from './components/modules/Payment/Payment';
 import Cart from './components/modules/Cart/Cart';
+import User from './components/modules/User/User';
+import UserProducts from './components/modules/User/userProductsPage/UserProducts';
+import UserCart from './components/modules/User/userCarts/UserCart';
 
 import { CartContextProvider } from './context/cart/cart';
 import ProtectedRoute from './ProtectedRoute';
