--- packages/frontend/src/components/containers/index.ts
@@ -1,6 +1,0 @@
-import PayPal from '../modules/Payment/PayPal';
-import CartContent from '../modules/Cart/CartContent';
-import UserForms from './UserForms';
-import UserProducts from './UserProducts';
-
-export { PayPal, CartContent, UserForms, UserProducts };
