+++ packages/frontend/src/components/molecules/index.ts
@@ -1,5 +1,5 @@
 import Searcher from './Searcher';
 import BasicInfo from './BasicInfo';
-import ListProducts from './ListProducts';
+import ListProducts from '../modules/User/ListProducts';
 
 export { Searcher, BasicInfo, ListProducts };
