+++ packages/frontend/src/components/pages/index.ts
@@ -1,6 +1,6 @@
 import Main from './Main';
-import User from './User';
-import UserProducts from './UserProducts';
-import UserCart from './UserCart';
+import User from '../modules/User/User';
+import UserProducts from '../modules/User/userProductsPage/UserProducts';
+import UserCart from '../modules/User/userCarts/UserCart';
 
 export { Main, User, UserProducts, UserCart };
